moved confirmation modal into the customPropertyField rather than the field array
