+++ lib/CustomPropertiesConfigListFieldArray/CustomPropertiesConfigListFieldArray.js
@@ -1,16 +1,14 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { isEqual } from 'lodash';
-import { defineMessages, injectIntl, FormattedMessage } from 'react-intl';
+import { FormattedMessage } from 'react-intl';
 import { Field } from 'react-final-form';
-import SafeHTMLMessage from '@folio/react-intl-safe-html';
 
-import { CalloutContext } from '@folio/stripes/core';
-import { Button, Col, ConfirmationModal, Row } from '@folio/stripes/components';
+import { Button, Col, Row } from '@folio/stripes/components';
 import CustomPropertyField from './CustomPropertyField';
 
 
-class CustomPropertiesConfigListFieldArray extends React.Component {
+export default class CustomPropertiesConfigListFieldArray extends React.Component {
   static propTypes = {
     fields: PropTypes.shape({
       name: PropTypes.string,
@@ -22,16 +20,10 @@ class CustomPropertiesConfigListFieldArray extends React.Component {
     onDelete: PropTypes.func.isRequired,
     onSave: PropTypes.func.isRequired,
     pickLists: PropTypes.arrayOf(PropTypes.object),
-    intl: PropTypes.shape({
-      formatMessage: PropTypes.func.isRequired,
-    }),
     translationKey: PropTypes.string,
   };
 
-  constructor(props) {
-    super(props);
-    this.state = { showConfirmDelete: false, disableNewButton: false, deleteIndex: 0, deleteName: '' };
-  }
+  state = { disableNewButton: false };
 
   defaultCustomPropertyValue = {
     defaultInternal: true,
@@ -66,24 +58,8 @@ class CustomPropertiesConfigListFieldArray extends React.Component {
     return this.props.onSave(customProperty);
   };
 
-  showDeleteConfirmationModal = () => this.setState({ showConfirmDelete: true });
-
-  hideDeleteConfirmationModal = () => this.setState({ showConfirmDelete: false });
-
   render() {
     const { fields, mutators, pickLists, translationKey } = this.props;
-    const { formatMessage } = this.props.intl;
-
-    const messages = defineMessages({
-      customProperty: {
-        id: `stripes-erm-components.customProperties.${translationKey}`,
-        defaultMessage: <FormattedMessage id="stripes-erm-components.customProperty" />,
-      },
-      customPropertyLowercase: {
-        id: `stripes-erm-components.customProperties.${translationKey}.lowercase`,
-        defaultMessage: <FormattedMessage id="stripes-erm-components.customProperty.lowercase" />,
-      },
-    });
 
     return (
       <div>
@@ -109,9 +85,7 @@ class CustomPropertiesConfigListFieldArray extends React.Component {
                 mutators={mutators}
                 name={`${fields.name}[${i}]`}
                 onDelete={() => {
-                  this.setState({ deleteIndex: i });
-                  this.setState({ deleteName: customProperty.name });
-                  this.showDeleteConfirmationModal();
+                  this.handleDelete(i);
                 }}
                 onSave={() => this.handleSave(i)}
                 pickLists={pickLists}
@@ -125,21 +99,7 @@ class CustomPropertiesConfigListFieldArray extends React.Component {
             </>
           );
         })}
-        {this.state.showConfirmDelete && (
-          <ConfirmationModal
-            id="delete-job-confirmation"
-            confirmLabel={<FormattedMessage id="stripes-erm-components.customProperties.delete.confirmLabel" />}
-            heading={<FormattedMessage id="stripes-erm-components.customProperties.delete.confirmHeading" values={{ custProp: formatMessage(messages.customPropertyLowercase) }} />}
-            message={<SafeHTMLMessage id="stripes-erm-components.customProperties.delete.confirmMessage" values={{ custProp: formatMessage(messages.customProperty), name: this.state.deleteName }} />}
-            onCancel={this.hideDeleteConfirmationModal}
-            onConfirm={this.handleDelete(this.state.deleteIndex)}
-            buttonStyle="danger"
-            open
-          />
-        )}
       </div>
     );
   }
 }
-
-export default injectIntl(CustomPropertiesConfigListFieldArray);
