+++ lib/CustomPropertiesConfigListFieldArray/CustomPropertyField.js
@@ -1,10 +1,15 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 
+import { defineMessages, injectIntl, FormattedMessage } from 'react-intl';
+import SafeHTMLMessage from '@folio/react-intl-safe-html';
+
+import { ConfirmationModal } from '@folio/stripes/components';
+
 import CustomPropertyFieldEdit from './CustomPropertyFieldEdit';
 import CustomPropertyFieldView from './CustomPropertyFieldView';
 
-export default class CustomPropertyField extends React.Component {
+class CustomPropertyField extends React.Component {
   static propTypes = {
     input: PropTypes.shape({
       name: PropTypes.string.isRequired,
@@ -13,11 +18,15 @@ export default class CustomPropertyField extends React.Component {
         PropTypes.string,
       ]).isRequired,
     }).isRequired,
+    intl: PropTypes.shape({
+      formatMessage: PropTypes.func.isRequired,
+    }),
     mutators: PropTypes.shape({
       setCustomPropertyValue: PropTypes.func.isRequired,
     }).isRequired,
     onDelete: PropTypes.func.isRequired,
     onSave: PropTypes.func.isRequired,
+    translationKey: PropTypes.string
   }
 
   constructor(props) {
@@ -28,6 +37,7 @@ export default class CustomPropertyField extends React.Component {
     this.state = {
       editing: !(value && value.id),
       initialValue: value,
+      showConfirmDelete: false,
     };
   }
 
@@ -61,17 +71,54 @@ export default class CustomPropertyField extends React.Component {
       .then(() => this.setState({ editing: false }));
   }
 
+  showDeleteConfirmationModal = () => this.setState({ showConfirmDelete: true });
+
+  hideDeleteConfirmationModal = () => this.setState({ showConfirmDelete: false });
+
   render() {
     const CustomPropertyComponent = this.state.editing ? CustomPropertyFieldEdit : CustomPropertyFieldView;
+    const { translationKey } = this.props;
+    const { formatMessage } = this.props.intl;
+    const custPropName = this.props?.input?.value?.name;
+
+    const messages = defineMessages({
+      customProperty: {
+        id: `stripes-erm-components.customProperties.${translationKey}`,
+        defaultMessage: <FormattedMessage id="stripes-erm-components.customProperty" />,
+      },
+      customPropertyLowercase: {
+        id: `stripes-erm-components.customProperties.${translationKey}.lowercase`,
+        defaultMessage: <FormattedMessage id="stripes-erm-components.customProperty.lowercase" />,
+      },
+    });
 
     return (
-      <CustomPropertyComponent
-        {...this.props}
-        onCancel={this.handleCancel}
-        onDelete={this.props.onDelete}
-        onSave={this.handleSave}
-        onEdit={this.handleEdit}
-      />
+      <>
+        <CustomPropertyComponent
+          {...this.props}
+          onCancel={this.handleCancel}
+          onDelete={this.showDeleteConfirmationModal}
+          onSave={this.handleSave}
+          onEdit={this.handleEdit}
+        />
+        {this.state.showConfirmDelete && (
+          <ConfirmationModal
+            id="delete-job-confirmation"
+            confirmLabel={<FormattedMessage id="stripes-erm-components.customProperties.delete.confirmLabel" />}
+            heading={<FormattedMessage id="stripes-erm-components.customProperties.delete.confirmHeading" values={{ custProp: formatMessage(messages.customPropertyLowercase) }} />}
+            message={<SafeHTMLMessage id="stripes-erm-components.customProperties.delete.confirmMessage" values={{ custProp: formatMessage(messages.customProperty), name: custPropName }} />}
+            onCancel={this.hideDeleteConfirmationModal}
+            onConfirm={() => {
+              this.props.onDelete();
+              this.hideDeleteConfirmationModal();
+            }}
+            buttonStyle="danger"
+            open
+          />
+        )}
+      </>
     );
   }
 }
+
+export default injectIntl(CustomPropertyField);
