+++ translations/stripes-erm-components/en.json
@@ -66,7 +66,7 @@
   "customPropertiesList.notSet": "Not set",
 
   "customProperties.delete.confirmLabel": "Delete",
-  "customProperties.delete.confirmHeading": "Delete {custprop}",
+  "customProperties.delete.confirmHeading": "Delete {custProp}",
   "customProperties.delete.confirmMessage": "{custProp} <strong>{name}</strong> will be <strong>deleted</strong>.",
